Fix terminology
Fixing the proper naming: old info was "TR01", but this is confusing, as the proper name for the component is "TROPIC01". If this would stay uncorrected - the assumption is that there maybe is an older revision or another component out there. TR01 seems to be internal naming (see ODU_TR01_user_api_v1.3.0.pdf) but confusing for external users.